pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "taskmanager"
version = "0.1.0"
description = "TaskManager API with auth"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.104.1",
  "starlette>=0.27,<0.38",
  "uvicorn==0.24.0",
  "pydantic>=2.3,<3",
  "pydantic[email]",
  "pydantic-settings==2.1.0",
  "python-multipart==0.0.6",
  "httpx",
  "sqlalchemy==2.0.23",
  "aiosqlite==0.19.0",
  "requests",
  "passlib[bcrypt]==1.7.4",
  "python-jose[cryptography]==3.3.0",
  "python-dotenv==1.0.1",
]

[project.scripts]
taskmanager = "app.scripts.run_app:main"

[dependency-groups]
dev = [
  "pytest>=8.2.0",
  "pytest-bdd",
]


